Commit for 4.0.4rc1 - Fix DMA buffer problem for ni_usb - Fix Bug: #60 type mismatch pc2_command patch from dac922 - Bumped version number to 4.0.4
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1656 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/drivers/gpib/pc2/pc2_init.c
@@ -49,7 +49,7 @@ int pc2_write(gpib_board_t *board, uint8_t *buffer, size_t length, int send_eoi,
 	pc2_private_t *priv = board->private_data;
 	return nec7210_write(board, &priv->nec7210_priv, buffer, length, send_eoi, bytes_written);
 }
-ssize_t pc2_command(gpib_board_t *board, uint8_t *buffer, size_t length, size_t *bytes_written)
+int pc2_command(gpib_board_t *board, uint8_t *buffer, size_t length, size_t *bytes_written)
 {
 	pc2_private_t *priv = board->private_data;
 	return nec7210_command(board, &priv->nec7210_priv, buffer, length, bytes_written);
